requirements.txt
streamlit>=1.32
altair>=5.0
pandas>=2.0
pyyaml>=6.0
requests>=2.31
openpyxl>=3.1
